Converts api to typescript
